Fix deadlock when using NSFileHandle:fileClose in mocked NSTask
When calling NSFileHandle:fileClose on the write end of stdout/stderr
pipes in a mocked NSTask (OCMocjObject+NSTask) a race condition could
lead to a deadlock This was because client code is using
NSFileHandle:readInBackgrond* and therefore spawned new threads for
reading which apparently increases a seaphore on the filedescriptor.
By moving those read-calls after the call to NSTask:launch the problem
is solved and we can use [NSFileHandle:fileClose] without risking
dead-lock situation.