Fix check for status of RAID device in leopard
In Leopard the output of 'diskutil info -plist <device-of-raidslice>' is
lacking a value for RAIDSetStatus-key. We circumvent this restriction by
checking the status of *every* raid set in the system (not only the one
of the source slice) and additionally gain some robustness by denying
operaition on a system with missing/failing/rebuilding drives.